requirements.txt
requests
python-dotenv
socketio
opencv-python
av